% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidata.R
\name{epidata}
\alias{epidata}
\title{Create an epivakt data set}
\usage{
epidata(id, date, location, organism, date_format = "\%Y.\%m.\%d")
}
\arguments{
\item{id}{Vector containing relevant identification of records in the data.
In analytix data this might be the TestId field}

\item{date}{Vector of strings containing dates. In analytix data dates are
provided by the RequisitionDate field}

\item{location}{Vector containing references to locations. In analytix data
the RequisitionCode field may be used as such}

\item{organism}{Vector containing references to bacteria found in samples. In
analytix data this will correspond to the FindingResult field}

\item{date_format}{Character string defining the format for \code{date} and must
correspond to whatever format provided by data. Default format is set to
\code{"\%Y.\%m.\%d"}}
}
\value{
A data frame with epivakt data
}
\description{
Function that creates an epivakt data set for further analysis. Input to this
function will be raw data from laboratory systems such as analytix.
Currently, this function only deal with the minimum set of variables and
it is likely to be extended.
}
\details{
Any mapping of variables performed within this function will use
mapping defined in static data sets provided by the epivakt R package. Hence,
changes in mappings will be part of updates of the epivakt R package.
}
\examples{
# Make some data
id <- 1:10
date <- rep("2022.10.31", 10)
location <- LETTERS[1:10]
organism <- letters[19:28]

# Make a data set
data.frame(epidata(id, date, location, organism))
}
