% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category.R
\name{missing_category}
\alias{missing_category}
\title{Print frequency of missing categories in data}
\usage{
missing_category(data, category)
}
\arguments{
\item{data}{Vector of raw, uncategorised data}

\item{category}{Vector of categorized values corresponding to \code{data} as
returned by \code{categorize()}}
}
\value{
A sorted frequency table of missing categories
}
\description{
Print frequency of missing categories in data
}
\examples{
data <- sample(c("gerA", "geA", "gerB", "germC", "gD"), 100, replace = TRUE)
mapping <- data.frame(
  name = c("gerA", "geA", "gerB"),
  value = c("GermA", "GermA", "GermB")
)

# Corresponding categories
category <- categorize(data, mapping)

# Occurrence of data with no matching category
missing_category(data, category)
}
