% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{organism_map}
\alias{organism_map}
\title{epivakt organism mapping data}
\format{
A data frame with 452 rows and 2 columns:
\describe{
\item{name}{Textual finding result found in raw laboratoty data}
\item{value}{Category to be used in further analysis}
}
}
\source{
WHONET software
}
\usage{
organism_map
}
\description{
A data set allowing for mapping of textual result findings in raw laboratory
data into known categories used in further analysis
}
\keyword{datasets}
