% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mapping.R
\name{read_mapping}
\alias{read_mapping}
\alias{read_analytix_mapping}
\alias{read_antibiotic_map}
\alias{read_specimen_map}
\alias{read_organism_map}
\title{Read name value mappings from file}
\usage{
read_mapping(file, type, source = "analytix")

read_analytix_mapping(file)

read_antibiotic_map(file, source = "analytix")

read_specimen_map(file, source = "analytix")

read_organism_map(file, source = "analytix")
}
\arguments{
\item{file}{Path and name of file containing the mapping}

\item{type}{What to map, one of
\code{c("antibiotic", "specimen", "organism")}}

\item{source}{Where does the mapping come from. Currently the only valid
option is \code{"analytix"}}
}
\value{
A data frame containing the \emph{value} that the corresponding \emph{name}
found in the data should be mapped
}
\description{
When using mappings obtained from WHONET input files will be named
"CodeDict_" with some further reference to content such as antibiotic,
organism or specimen.
}
